Consolidate MemoryString to use the lib conversion functions
